requirements.txt
transparent-background
Pillow
